Header Updated
I had an idea. I really like the idea of keeping the icons for the time
being, so I figured we could swap out the Email icon in the mobile
header to a Submit one. Add Submit to the desktop nav and keep the
icons in the mobile nav until we get more pages. We have the sticky
contact us button on the right and the contact form on the bottom. What
do you think?